+++ benchmarks/appsec_iast_aspects/scenario.py
@@ -6,8 +6,8 @@ from bm.iast_utils import asm_config
 from bm.utils import override_env
 
 from ddtrace.appsec._iast import enable_iast_propagation
-from ddtrace.appsec._iast._iast_request_context_base import _num_objects_tainted_in_request
 from ddtrace.appsec._iast._taint_tracking import OriginType
+from ddtrace.appsec._iast._taint_tracking._context import _num_objects_tainted_in_request
 from ddtrace.appsec._iast._taint_tracking._context import debug_context_array_free_slots_number
 from ddtrace.appsec._iast._taint_tracking._context import debug_context_array_size
 from ddtrace.appsec._iast._taint_tracking._taint_objects import taint_pyobject
