+++ tests/appsec/iast/test_safe_wrappers_and_initialization.py
@@ -116,13 +116,13 @@ class TestNativeStateInitialization:
 
     def test_taint_operations_work_after_initialization(self):
         """Test that taint operations work correctly after initialization."""
-        from ddtrace.appsec._iast._iast_request_context_base import IAST_CONTEXT
         from ddtrace.appsec._iast._overhead_control_engine import oce
         from ddtrace.appsec._iast._taint_tracking import initialize_native_state
         from ddtrace.appsec._iast._taint_tracking import reset_native_state
         from ddtrace.appsec._iast._taint_tracking._context import start_request_context
         from ddtrace.appsec._iast._taint_tracking._taint_objects import taint_pyobject
         from ddtrace.appsec._iast._taint_tracking._taint_objects_base import is_pyobject_tainted
+        from ddtrace.appsec._shared._iast_context import IAST_CONTEXT
         from ddtrace.internal.settings.asm import config as asm_config
         from tests.utils import override_env
